THRIFT-5185: Renamed endian.h. Fix memory leak. Silenced warnings. Client: cpp
