+++ lib/cpp/Makefile.am
@@ -142,7 +142,7 @@ include_thrift_HEADERS = \
                          src/thrift/TLogging.h \
                          src/thrift/TToString.h \
                          src/thrift/TBase.h \
-                         src/thrift/endian.h
+                         src/thrift/portable_endian.h
 
 include_concurrencydir = $(include_thriftdir)/concurrency
 include_concurrency_HEADERS = \
